File order is preserved for javascript files -- css files are sorted.
This is because there might be deps within the javascript files, so preserving
the order in which they were added to a group is important.
However, no such deps should exist in css files (and the spec doesn't say
anything about which has precidence if there are two styles with identical
selectors), so we can safely allow re-ordering.