Remove declaration of 'global' group.
The upshot of this is that the global group is now created when a file
is first added to it, rather than at the beginning of the program.
Since groups are rendered in the order in which they appear in the
groups array, this means that non-global groups can be rendered first,
which might be useful for satisfying dependancies.