Add support for remote file paths. The path is specified in 'paths', like so: 'paths => array(    'core' => 'assets/',    'remote' => 'http://localhost/whatever/assets/', )
Then, when that path is used when specifying a file, that remote location is checked.
Note that js_dir, css_dir, img_dir are **NOTE** used here, as it may be you're
referring to an external url outside of your control.
(At some point in the future I may make js_dir, etc, per-path, but not yet).
Also, globbing won't work (for obvious reasons), and caching will assume
the file is unchanged (as we don't want a ton of requests going off to the remote server).