Add gen_tags option to render_js and render_css.
This patch adds a second argument to render_js and render_css called
$options. This can be expended in future, but for now you can specify
a single key, 'gen_tags'.

With this set to false, render_* will return an array of the filenames /
file contents (depending on the inline setting), instead of generating
the tags itself.

Backwards compatibility with the old second argument to render_*, $inline,
is retained.

Closes #14